Restructure the whole api project
