+++ api/.gitignore
@@ -1,5 +1,4 @@
 dist/
-data/
 history.txt
 public/images/*
 public/uploads/*
