+++ api/src/data/db.ts
@@ -1,7 +1,8 @@
-/**
- * Build PostgreSQL connection URL from component env vars.
- * Required: DB_USER, DB_PASS, DB_PORT, DB_DOMAIN, DB_EVY_DATABASE.
- */
+import { drizzle } from "drizzle-orm/postgres-js";
+import postgres from "postgres";
+
+import * as schema from "../../../types/generated/ts/db/schema.generated";
+
 export function getConnectionUrl(): string {
 	const user = process.env.DB_USER;
 	const pass = process.env.DB_PASS;
@@ -26,3 +27,26 @@ export function getConnectionUrl(): string {
 	const encodedPass = encodeURIComponent(pass);
 	return `postgresql://${encodedUser}:${encodedPass}@${domain}:${port}/${database}`;
 }
+
+const connectionString = getConnectionUrl();
+const client = postgres(connectionString);
+
+let db = drizzle(client, { schema });
+
+export function setDbForTest(database: typeof db): void {
+	db = database;
+}
+
+export function hasDatabaseErrorCode(err: unknown, code: string): boolean {
+	if (typeof err !== "object" || err === null) {
+		return false;
+	}
+
+	if ("code" in err && err.code === code) {
+		return true;
+	}
+
+	return "cause" in err && hasDatabaseErrorCode(err.cause, code);
+}
+
+export { db };
