+++ api/src/readiness.ts
@@ -2,7 +2,7 @@
 // so Docker/e2e use this CLI to
 // verify DB and seed readiness.
 import type { GetRequest, GetResponse } from "evy-types";
-import { get as defaultGet } from "./rpc";
+import { get as defaultGet } from "./procedures/rpc";
 
 type AssertApiReadableOptions = {
 	requireSeeded: boolean;
